♻️ refactor: addData() 파일 분리
