+++ src/addData.js
@@ -1,0 +1,17 @@
+import { categoryValue } from "./optionValue.js";
+import { RestaurantData } from "./RestaurantData.js";
+
+const addData = () => {
+  const formData = new FormData(document.getElementById("input-form"));
+  const submittedData = Object.fromEntries(formData);
+  const information = {
+    name: submittedData.name,
+    distance: Number(submittedData.distance),
+    description: submittedData.description,
+    imgSrc: `../templates/category-${submittedData.category}.png`,
+    imgAlt: `${categoryValue[submittedData.category]}`,
+  };
+  RestaurantData.push(information);
+};
+
+export default addData;
